requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.94.1
h11==0.14.0
idna==3.4
pydantic==1.10.6
sniffio==1.3.0
starlette==0.26.1
typing-extensions==4.5.0
uvicorn==0.21.1
python-multipart==0.0.6
python-dotenv==0.19.0
dnspython==2.3.0
motor==2.5.0


pymongo~=3.13.0